+++ examples/mongodbatlas_flex_cluster/versions.tf
@@ -1,8 +1,7 @@
 terraform {
   required_providers {
     mongodbatlas = {
-      source  = "mongodb/mongodbatlas"
-      version = "~> 1.21.3"
+      source = "mongodb/mongodbatlas"
     }
   }
   required_version = ">= 1.0"