+++ internal/service/federatedsettingsorgconfig/model_federated_settings_connected_orgs.go
@@ -4,7 +4,7 @@ import (
 	"sort"
 	"strings"
 
-	"go.mongodb.org/atlas-sdk/v20250312005/admin"
+	"go.mongodb.org/atlas-sdk/v20250312006/admin"
 )
 
 type roleMappingsByGroupName []admin.AuthFederationRoleMapping