+++ internal/service/organization/resource_organization_test.go
@@ -9,7 +9,7 @@ import (
 	"strings"
 	"testing"
 
-	"go.mongodb.org/atlas-sdk/v20250312005/admin"
+	"go.mongodb.org/atlas-sdk/v20250312006/admin"
 
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 	"github.com/hashicorp/terraform-plugin-testing/terraform"
@@ -220,10 +220,11 @@ func TestAccConfigRSOrganization_import(t *testing.T) {
 				Config: configImportSet(orgID, orgName), // Use import so a new organization is not created, the resource must exist in a step before import state is verified.
 			},
 			{
-				ResourceName:      resourceName,
-				ImportStateId:     orgID,
-				ImportState:       true, // Do the import check.
-				ImportStateVerify: true,
+				ResourceName:                         resourceName,
+				ImportStateId:                        orgID,
+				ImportState:                          true, // Do the import check.
+				ImportStateVerify:                    true,
+				ImportStateVerifyIdentifierAttribute: "org_id",
 			},
 			{
 				// Use removed block so the organization is not deleted.