+++ internal/testutil/acc/shared_resource.go
@@ -1,6 +1,7 @@
 package acc
 
 import (
+	"context"
 	"fmt"
 	"os"
 	"sync"
@@ -8,11 +9,13 @@ import (
 	"time"
 
 	"github.com/mongodb/terraform-provider-mongodbatlas/internal/config"
+	"github.com/mongodb/terraform-provider-mongodbatlas/internal/testutil/clean"
 	"github.com/stretchr/testify/require"
 )
 
 const (
 	MaxClusterNodesPerProject = 30 // Choose to be conservative, 40 clusters per project is the limit before `CROSS_REGION_NETWORK_PERMISSIONS_LIMIT_EXCEEDED` error, see https://www.mongodb.com/docs/atlas/reference/atlas-limits/
+	MaxFreeTierClusterCount   = 1  // Project can have at most 1 free tier cluster
 )
 
 // SetupSharedResources must be called from TestMain test package in order to use ProjectIDExecution.
@@ -32,7 +35,16 @@ func cleanupSharedResources() {
 		fmt.Printf("Deleting execution cluster: %s, project id: %s\n", sharedInfo.clusterName, projectID)
 		deleteCluster(projectID, sharedInfo.clusterName)
 	}
-
+	if sharedInfo.streamInstanceName != "" {
+		projectID := sharedInfo.projectID
+		if projectID == "" {
+			projectID = projectIDLocal() // Maybe we are using an existing project, but stream instance is not used
+		}
+		_, err := clean.RemoveStreamInstances(context.TODO(), false, ConnV2(), projectID)
+		if err != nil {
+			fmt.Printf("Failed to delete stream instances: for execution project %s, error: %s\n", projectID, err)
+		}
+	}
 	if sharedInfo.projectID != "" {
 		fmt.Printf("Deleting execution project: %s, id: %s\n", sharedInfo.projectName, sharedInfo.projectID)
 		deleteProject(sharedInfo.projectID)
@@ -68,11 +80,8 @@ func ProjectIDExecution(tb testing.TB) string {
 	return sharedInfo.projectID
 }
 
-// ProjectIDExecutionWithCluster creates a project and reuses it for  `MaxClusterNodesPerProject ` nodes. The clusterName is always unique.
-// TotalNodeCount = sum(specs.node_count) * num_shards (1 if new schema)
-// This avoids the `CROSS_REGION_NETWORK_PERMISSIONS_LIMIT_EXCEEDED` error when creating too many clusters within the same project.
-// When `MONGODB_ATLAS_PROJECT_ID` and `MONGODB_ATLAS_CLUSTER_NAME` are defined, they are used instead of creating a project and clusterName.
-func ProjectIDExecutionWithCluster(tb testing.TB, totalNodeCount int) (projectID, clusterName string) {
+// ProjectIDExecutionWithFreeCluster is identical to ProjectIDExecutionWithCluster but also contemplates the restriction of `MaxFreeTierClusterCount`
+func ProjectIDExecutionWithFreeCluster(tb testing.TB, totalNodeCount, freeTierClusterCount int) (projectID, clusterName string) {
 	tb.Helper()
 	if ExistingClusterUsed() {
 		return existingProjectIDClusterName()
@@ -80,9 +89,18 @@ func ProjectIDExecutionWithCluster(tb testing.TB, totalNodeCount int) (projectID
 	// Only skip after ExistingClusterUsed() to allow MacT (Mocked-Acceptance Tests) to return early instead of being skipped.
 	SkipInUnitTest(tb)
 	require.True(tb, sharedInfo.init, "SetupSharedResources must called from TestMain test package")
-	return NextProjectIDClusterName(totalNodeCount, func(projectName string) string {
+	return NextProjectIDClusterName(totalNodeCount, freeTierClusterCount, func(projectName string) string {
 		return createProject(tb, projectName)
 	})
+}
+
+// ProjectIDExecutionWithCluster creates a project and reuses it with other tests respecting `MaxClusterNodesPerProject` restrictions. The clusterName is always unique.
+// TotalNodeCount = sum(specs.node_count) * num_shards (1 if new schema)
+// This avoids `CROSS_REGION_NETWORK_PERMISSIONS_LIMIT_EXCEEDED` and `project has reached the limit for the number of free clusters` errors when creating too many clusters within the same project.
+// When `MONGODB_ATLAS_PROJECT_ID` and `MONGODB_ATLAS_CLUSTER_NAME` are defined, they are used instead of creating a project and clusterName.
+func ProjectIDExecutionWithCluster(tb testing.TB, totalNodeCount int) (projectID, clusterName string) {
+	tb.Helper()
+	return ProjectIDExecutionWithFreeCluster(tb, totalNodeCount, 0)
 }
 
 // ClusterNameExecution returns the name of a created cluster for the execution of the tests in the resource package.
@@ -120,6 +138,32 @@ func ClusterNameExecution(tb testing.TB, populateSampleData bool) (projectID, cl
 	return projectID, sharedInfo.clusterName
 }
 
+// ProjectIDExecutionWithStreamInstance returns the project ID and stream instance name for test execution.
+// Uses the same ProjectID as the ProjectIDExecution.
+// The stream instance will include the `sample_stream_solar` connection. It is included to avoid ALREADY_EXIST errors as many different tests depends on this stream connection. Use a data source whenever you need it.
+// You can use `MONGODB_ATLAS_STREAM_INSTANCE_NAME` to use an "externally" managed stream instance.
+// We reuse a SPI to reduce the resource allocation (specially relevant for cloud dev).
+func ProjectIDExecutionWithStreamInstance(tb testing.TB) (projectID, streamInstanceName string) {
+	tb.Helper()
+	SkipInUnitTest(tb)
+	require.True(tb, sharedInfo.init, "SetupSharedResources must called from TestMain test package")
+	projectID = ProjectIDExecution(tb)
+
+	if existingStreamInstanceUsed() {
+		return projectID, existingStreamInstanceName()
+	}
+	sharedInfo.mu.Lock()
+	defer sharedInfo.mu.Unlock()
+	if sharedInfo.streamInstanceName == "" {
+		name := RandomStreamInstanceName()
+		tb.Logf("Creating execution stream instance: %s\n", name)
+		sharedInfo.streamInstanceName = name
+		createStreamInstance(tb, projectID, name)
+	}
+
+	return projectID, sharedInfo.streamInstanceName
+}
+
 // SerialSleep waits a few seconds so clusters in a project are not created concurrently, see HELP-65223.
 // This must be called once the test is marked as parallel, e.g. in PreCheck inside Terraform tests.
 func SerialSleep(tb testing.TB) {
@@ -134,29 +178,33 @@ func SerialSleep(tb testing.TB) {
 }
 
 type projectInfo struct {
-	id        string
-	name      string
-	nodeCount int
+	id                   string
+	name                 string
+	nodeCount            int
+	freeTierClusterCount int
 }
 
 var sharedInfo = struct {
-	projectID   string
-	projectName string
-	clusterName string
-	projects    []projectInfo
-	mu          sync.Mutex
-	muSleep     sync.Mutex
-	init        bool
+	projectID          string
+	projectName        string
+	clusterName        string
+	streamInstanceName string
+	projects           []projectInfo
+	mu                 sync.Mutex
+	muSleep            sync.Mutex
+	init               bool
 }{
 	projects: []projectInfo{},
 }
 
-// NextProjectIDClusterName is an internal method used when we want to reuse a projectID `MaxClustersPerProject` times
-func NextProjectIDClusterName(totalNodeCount int, projectCreator func(string) string) (projectID, clusterName string) {
+// NextProjectIDClusterName is an internal method used when we want to reuse a projectID respecting `MaxClustersNodesPerProject` and `MaxFreeTierClusterCount`
+func NextProjectIDClusterName(totalNodeCount, freeTierClusterCount int, projectCreator func(string) string) (projectID, clusterName string) {
 	sharedInfo.mu.Lock()
 	defer sharedInfo.mu.Unlock()
 	var project projectInfo
-	if len(sharedInfo.projects) == 0 || sharedInfo.projects[len(sharedInfo.projects)-1].nodeCount+totalNodeCount > MaxClusterNodesPerProject {
+	if len(sharedInfo.projects) == 0 ||
+		sharedInfo.projects[len(sharedInfo.projects)-1].nodeCount+totalNodeCount > MaxClusterNodesPerProject ||
+		sharedInfo.projects[len(sharedInfo.projects)-1].freeTierClusterCount+freeTierClusterCount > MaxFreeTierClusterCount {
 		project = projectInfo{
 			name:      RandomProjectName(),
 			nodeCount: totalNodeCount,
@@ -166,6 +214,7 @@ func NextProjectIDClusterName(totalNodeCount int, projectCreator func(string) st
 	} else {
 		project = sharedInfo.projects[len(sharedInfo.projects)-1]
 		sharedInfo.projects[len(sharedInfo.projects)-1].nodeCount += totalNodeCount
+		sharedInfo.projects[len(sharedInfo.projects)-1].freeTierClusterCount += freeTierClusterCount
 	}
 	return project.id, RandomClusterName()
 }